refactor: 토큰세팅 메서드 분리
Related to : #20
